requirements.txt
simple-slurm
kafka-python
psutil>=5.6.6
python-math
faust-streaming
werkzeug
